Recipe builder update, pt. 2
Allow library users to make their own scuffed recipe builder implementations. Useful for things like malum
